move misc. algorithm headers into algo subfolder
