+++ benchmark/benchmark_bedrock_channel.cpp
@@ -11,7 +11,7 @@
 
 #include "fastscapelib/grid/raster_grid.hpp"
 #include "fastscapelib/flow_graph.hpp"
-#include "fastscapelib/flow_routing.hpp"
+#include "fastscapelib/algo/flow_routing.hpp"
 #include "fastscapelib/bedrock_channel.hpp"
 
 #include "benchmark_setup.hpp"
