+++ include/fastscapelib/fastscapelib.hpp
@@ -9,7 +9,7 @@
 
 #include "fastscapelib/version.hpp"
 #include "fastscapelib/bedrock_channel.hpp"
-#include "fastscapelib/flow_routing.hpp"
+#include "fastscapelib/algo/flow_routing.hpp"
 #include "fastscapelib/flow_graph.hpp"
 #include "fastscapelib/flow_router.hpp"
 #include "fastscapelib/sink_resolver.hpp"
@@ -19,6 +19,6 @@
 #include "fastscapelib/grid/profile_grid.hpp"
 #include "fastscapelib/grid/raster_grid.hpp"
 #include "fastscapelib/hillslope.hpp"
-#include "fastscapelib/sinks.hpp"
+#include "fastscapelib/algo/pflood.hpp"
 
 #endif
