+++ include/fastscapelib/flow_router.hpp
@@ -9,7 +9,7 @@
 #ifndef FASTSCAPELIB_FLOW_ROUTER_H
 #define FASTSCAPELIB_FLOW_ROUTER_H
 
-#include "fastscapelib/flow_routing.hpp"
+#include "fastscapelib/algo/flow_routing.hpp"
 #include "fastscapelib/utils/xtensor_utils.hpp"
