+++ python/src/modelings.cpp
@@ -3,7 +3,7 @@
 
 #include "fastscapelib/bedrock_channel.hpp"
 #include "fastscapelib/hillslope.hpp"
-#include "fastscapelib/flow_routing.hpp"
+#include "fastscapelib/algo/flow_routing.hpp"
 #include "fastscapelib/grid/raster_grid.hpp"
 
 #include "xtensor-python/pytensor.hpp"
